requirements.txt
Django==4.2.16
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
coverage==7.6.1
gunicorn==23.0.0
